Make FakeOrcDataSource a top level class
